+++ empire/test/conftest.py
@@ -2,6 +2,7 @@ import asyncio
 import os
 import shutil
 import sys
+import warnings
 from contextlib import contextmanager, suppress
 from pathlib import Path
 from typing import TYPE_CHECKING
@@ -139,9 +140,10 @@ def regular_auth_token(client, admin_auth_token):
     return response.json()["access_token"]
 
 
-# TODO: Remove this. Use session_local
+# TODO: Remove this. Use session_local fixture instead
 @pytest.fixture(scope="module")
 def db():
+    warnings.warn("Use session_local fixture instead", DeprecationWarning, stacklevel=5)
     from empire.server.core.db.base import SessionLocal
 
     return SessionLocal()